+++ Sources/RxSwift/AppRouter+reactive.swift
@@ -10,71 +10,39 @@ import Foundation
 import UIKit
 import RxSwift
 
-public protocol ARReactiveProxyProtocol {}
-extension UIViewController : ARReactiveProxyProtocol {}
-public extension ARReactiveProxyProtocol where Self: UIViewController {
+public extension Reactive where Base: UIViewController {
     /// Observe viewDidLoad calls on current instance
     public func onViewDidLoad() -> Observable<Void> {
-        return Observable.create({ observer -> Disposable in
-            return ARViewControllerLifeCircleManager.instance.didLoad.subscribe(onNext: { [weak self] vc in
-                if vc === self {
-                    observer.onNext()
-                }
-            })
-        })
+        return ARViewControllerLifeCircleManager.instance.didLoad.filter{ [weak base] in $0 === base }.map{ _ in () }
     }
     
     /// Observe viewWillAppear calls on current instance
     public func onViewWillAppear() -> Observable<Bool> {
-        return Observable.create({ observer -> Disposable in
-            return ARViewControllerLifeCircleManager.instance.willAppear.subscribe(onNext: { [weak self] (vc, animated) in
-                if vc === self {
-                    observer.onNext(animated)
-                }
-            })
-        })
+        return ARViewControllerLifeCircleManager.instance.willAppear.filter{ [weak base] in $0.controller === base }.map{ $0.animated }
     }
     
     /// Observe viewDidAppear calls on current instance
     public func onViewDidAppear() -> Observable<Bool> {
-        return Observable.create({ observer -> Disposable in
-            return ARViewControllerLifeCircleManager.instance.didAppear.subscribe(onNext: { [weak self] (vc, animated) in
-                if vc === self {
-                    observer.onNext(animated)
-                }
-            })
-        })
+        return ARViewControllerLifeCircleManager.instance.didAppear.filter{ [weak base] in $0.controller === base }.map{ $0.animated }
     }
     
     /// Observe viewWillDisappear calls on current instance
     public func onViewWillDisappear() -> Observable<Bool> {
-        return Observable.create({ observer -> Disposable in
-            return ARViewControllerLifeCircleManager.instance.willDisappear.subscribe(onNext: { [weak self] (vc, animated) in
-                if vc === self {
-                    observer.onNext(animated)
-                }
-            })
-        })
+        return ARViewControllerLifeCircleManager.instance.willDisappear.filter{ [weak base] in $0.controller === base }.map{ $0.animated }
     }
     
     /// Observe viewDidDisappear calls on current instance
     public func onViewDidDisappear() -> Observable<Bool> {
-        return Observable.create({ observer -> Disposable in
-            return ARViewControllerLifeCircleManager.instance.didDisappear.subscribe(onNext: { [weak self] (vc, animated) in
-                if vc === self {
-                    observer.onNext(animated)
-                }
-            })
-        })
+        return ARViewControllerLifeCircleManager.instance.didDisappear.filter{ [weak base] in $0.controller === base }.map{ $0.animated }
     }
 }
 
-public extension ARReactiveProxyProtocol where Self: UIViewController {
+public extension Reactive where Base: UIViewController {
     /// observe viewDidLoad calls on all instances of current type
-    public static func onViewDidLoad() -> Observable<Self> {
+    public static func onViewDidLoad() -> Observable<Base> {
         return Observable.create({ observer -> Disposable in
             return ARViewControllerLifeCircleManager.instance.didLoad.subscribe(onNext: { vc in
-                if let required = vc as? Self {
+                if let required = vc as? Base {
                     observer.onNext(required)
                 }
             })
@@ -82,10 +50,10 @@ public extension ARReactiveProxyProtocol where Self: UIViewController {
     }
     
     /// observe viewWillAppear calls on all instances of current type
-    public static func onViewWillAppear() -> Observable<(controller: Self, animated: Bool)> {
+    public static func onViewWillAppear() -> Observable<(controller: Base, animated: Bool)> {
         return Observable.create({ observer -> Disposable in
             return ARViewControllerLifeCircleManager.instance.willAppear.subscribe(onNext: { (vc, animated) in
-                if let required = vc as? Self {
+                if let required = vc as? Base {
                     observer.onNext((required, animated))
                 }
             })
@@ -93,10 +61,10 @@ public extension ARReactiveProxyProtocol where Self: UIViewController {
     }
     
     /// observe viewDidAppear calls on all instances of current type
-    public static func onViewDidAppear() -> Observable<(controller: Self, animated: Bool)> {
+    public static func onViewDidAppear() -> Observable<(controller: Base, animated: Bool)> {
         return Observable.create({ observer -> Disposable in
             return ARViewControllerLifeCircleManager.instance.didAppear.subscribe(onNext: { (vc, animated) in
-                if let required = vc as? Self {
+                if let required = vc as? Base {
                     observer.onNext((required, animated))
                 }
             })
@@ -104,10 +72,10 @@ public extension ARReactiveProxyProtocol where Self: UIViewController {
     }
     
     /// observe viewWillDisappear calls on all instances of current type
-    public static func onViewWillDisappear() -> Observable<(controller: Self, animated: Bool)> {
+    public static func onViewWillDisappear() -> Observable<(controller: Base, animated: Bool)> {
         return Observable.create({ observer -> Disposable in
             return ARViewControllerLifeCircleManager.instance.willDisappear.subscribe(onNext: { (vc, animated) in
-                if let required = vc as? Self {
+                if let required = vc as? Base {
                     observer.onNext((required, animated))
                 }
             })
@@ -115,10 +83,10 @@ public extension ARReactiveProxyProtocol where Self: UIViewController {
     }
     
     /// observe viewDidDisappear calls on all instances of current type
-    public static func onViewDidDisappear() -> Observable<(controller: Self, animated: Bool)> {
+    public static func onViewDidDisappear() -> Observable<(controller: Base, animated: Bool)> {
         return Observable.create({ observer -> Disposable in
             return ARViewControllerLifeCircleManager.instance.didDisappear.subscribe(onNext: { (vc, animated) in
-                if let required = vc as? Self {
+                if let required = vc as? Base {
                     observer.onNext((required, animated))
                 }
             })
@@ -128,14 +96,14 @@ public extension ARReactiveProxyProtocol where Self: UIViewController {
 
 private class ARViewControllerLifeCircleManager {
     private static var __once: () = {
-            swizzleUIViewControllerDidLoad()
-            swizzleUIViewControllerWillAppear()
-            swizzleUIViewControllerDidAppear()
-            swizzleUIViewControllerWillDisappear()
-            swizzleUIViewControllerDidDisappear()
-        }()
+        swapMethods(#selector(UIViewController.viewDidLoad), swizzled: #selector(UIViewController.ar_viewDidLoad))
+        swapMethods(#selector(UIViewController.viewWillAppear(_:)), swizzled: #selector(UIViewController.ar_viewWillAppear(_:)))
+        swapMethods(#selector(UIViewController.viewDidAppear(_:)), swizzled: #selector(UIViewController.ar_viewDidAppear(_:)))
+        swapMethods(#selector(UIViewController.viewWillDisappear(_:)), swizzled: #selector(UIViewController.ar_viewWillDisappear(_:)))
+        swapMethods(#selector(UIViewController.viewDidDisappear(_:)), swizzled: #selector(UIViewController.ar_viewDidDisappear(_:)))
+    }()
     static let instance : ARViewControllerLifeCircleManager = {
-        ARViewControllerLifeCircleManager.swizzleUIViewController()
+        _ = __once
         return ARViewControllerLifeCircleManager()
     }()
     fileprivate let didLoad = PublishSubject<UIViewController>()
@@ -143,42 +111,7 @@ private class ARViewControllerLifeCircleManager {
     fileprivate let didAppear = PublishSubject<(controller: UIViewController,animated: Bool)>()
     fileprivate let willDisappear = PublishSubject<(controller: UIViewController,animated: Bool)>()
     fileprivate let didDisappear = PublishSubject<(controller: UIViewController,animated: Bool)>()
-    
-    class func swizzleUIViewController() {
-        struct Static { static var token: Int = 0 }
-        _ = ARViewControllerLifeCircleManager.__once
-    }
-    
-    fileprivate class func swizzleUIViewControllerDidLoad() {
-        let originalSelector = #selector(UIViewController.viewDidLoad)
-        let swizzledSelector = #selector(UIViewController.ar_viewDidLoad)
-        swapMethods(originalSelector, swizzled: swizzledSelector)
-    }
-    
-    fileprivate class func swizzleUIViewControllerWillAppear() {
-        let originalSelector = #selector(UIViewController.viewWillAppear(_:))
-        let swizzledSelector = #selector(UIViewController.ar_viewWillAppear(_:))
-        swapMethods(originalSelector, swizzled: swizzledSelector)
-    }
-    
-    fileprivate class func swizzleUIViewControllerDidAppear() {
-        let originalSelector = #selector(UIViewController.viewDidAppear(_:))
-        let swizzledSelector = #selector(UIViewController.ar_viewDidAppear(_:))
-        swapMethods(originalSelector, swizzled: swizzledSelector)
-    }
-    
-    fileprivate class func swizzleUIViewControllerWillDisappear() {
-        let originalSelector = #selector(UIViewController.viewWillDisappear(_:))
-        let swizzledSelector = #selector(UIViewController.ar_viewWillDisappear(_:))
-        swapMethods(originalSelector, swizzled: swizzledSelector)
-    }
-    
-    fileprivate class func swizzleUIViewControllerDidDisappear() {
-        let originalSelector = #selector(UIViewController.viewDidDisappear(_:))
-        let swizzledSelector = #selector(UIViewController.ar_viewDidDisappear(_:))
-        swapMethods(originalSelector, swizzled: swizzledSelector)
-    }
-    
+
     fileprivate class func swapMethods(_ original: Selector, swizzled: Selector) {
         let originalMethod = class_getInstanceMethod(UIViewController.self, original)
         let swizzledMethod = class_getInstanceMethod(UIViewController.self, swizzled)
@@ -210,4 +143,3 @@ extension UIViewController {
         ARViewControllerLifeCircleManager.instance.didDisappear.onNext((self, animated: animated))
     }
 }
-